requirements.txt.txt
python-telegram-bot==21.6
Pillow==12.1.1
arabic-reshaper==3.0.0
python-bidi==0.4.2